aws: refactor aws.Config URL parsing to aws package
